% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_dash}
\alias{da_dash}
\title{da_dash}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 300361 rows and 17 columns.
}
\usage{
da_dash
}
\description{
da_dash
}
\keyword{datasets}
