% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_rfb}
\alias{da_rfb}
\title{da_rfb}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6967 rows and 31 columns.
}
\usage{
da_rfb
}
\description{
da_rfb
}
\keyword{datasets}
