% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_rfb_ativo}
\alias{da_rfb_ativo}
\title{da_rfb_ativo}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1657 rows and 31 columns.
}
\usage{
da_rfb_ativo
}
\description{
da_rfb_ativo
}
\keyword{datasets}
