% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_tempo_trf1}
\alias{da_tempo_trf1}
\title{da_tempo_trf1}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8212 rows and 6 columns.
}
\usage{
da_tempo_trf1
}
\description{
da_tempo_trf1
}
\keyword{datasets}
