% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_trf1_cjpg_pequeno}
\alias{da_trf1_cjpg_pequeno}
\title{da_trf1_cjpg_pequeno}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 227 rows and 14 columns.
}
\usage{
da_trf1_cjpg_pequeno
}
\description{
da_trf1_cjpg_pequeno
}
\keyword{datasets}
