% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_trf1_cpopg}
\alias{da_trf1_cpopg}
\title{da_trf1_cpopg}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10241 rows and 15 columns.
}
\usage{
da_trf1_cpopg
}
\description{
da_trf1_cpopg
}
\keyword{datasets}
