% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-app.R
\name{pegar_tempo}
\alias{pegar_tempo}
\title{Pegar tempo a partir do tidy kaplan meier}
\usage{
pegar_tempo(d_km, quantile, var = FALSE)
}
\arguments{
\item{d_km}{Tidy do modelo kaplan meier}

\item{quantile}{quantil}

\item{var}{agrupou por variável? Default \code{FALSE}}
}
\description{
Pegar tempo a partir do tidy kaplan meier
}
