% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{sf_amazon}
\alias{sf_amazon}
\title{sf_amazon}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 1 columns.
}
\usage{
sf_amazon
}
\description{
sf_amazon
}
\keyword{datasets}
