% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{sinesp}
\alias{sinesp}
\title{sinesp}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5466 rows and 3 columns.
}
\usage{
sinesp
}
\description{
sinesp
}
\keyword{datasets}
